Skeleton, waiting for Svelte 5
